+++ backend/src/apiserver/storage/experiment_store_test.go
@@ -57,7 +57,7 @@ func TestListExperiments_Pagination(t *testing.T) {
 	assert.Nil(t, err)
 	assert.NotEmpty(t, nextPageToken)
 	assert.Equal(t, experimentsExpected, experiments)
-	assert.Equal(t, 4, count)
+	assert.Equal(t, int32(4), count)
 
 	expectedExperiment2 := &model.Experiment{
 		UUID:           fakeIDTwo,
@@ -79,7 +79,7 @@ func TestListExperiments_Pagination(t *testing.T) {
 	experiments, count, nextPageToken, err = experimentStore.ListExperiments(opts)
 	assert.Nil(t, err)
 	assert.Empty(t, nextPageToken)
-	assert.Equal(t, 4, count)
+	assert.Equal(t, int32(4), count)
 	assert.Equal(t, experimentsExpected2, experiments)
 }
 
@@ -115,7 +115,7 @@ func TestListExperiments_Pagination_Descend(t *testing.T) {
 
 	assert.Nil(t, err)
 	assert.NotEmpty(t, nextPageToken)
-	assert.Equal(t, 4, count)
+	assert.Equal(t, int32(4), count)
 	assert.Equal(t, experimentsExpected, experiments)
 
 	expectedExperiment1 := &model.Experiment{
@@ -138,7 +138,7 @@ func TestListExperiments_Pagination_Descend(t *testing.T) {
 	experiments, count, nextPageToken, err = experimentStore.ListExperiments(opts)
 	assert.Nil(t, err)
 	assert.Empty(t, nextPageToken)
-	assert.Equal(t, 4, count)
+	assert.Equal(t, int32(4), count)
 	assert.Equal(t, experimentsExpected2, experiments)
 }
 
@@ -161,7 +161,7 @@ func TestListExperiments_Pagination_LessThanPageSize(t *testing.T) {
 	experiments, count, nextPageToken, err := experimentStore.ListExperiments(opts)
 	assert.Nil(t, err)
 	assert.Equal(t, "", nextPageToken)
-	assert.Equal(t, 1, count)
+	assert.Equal(t, int32(1), count)
 	assert.Equal(t, experimentsExpected, experiments)
 }
 
@@ -348,7 +348,7 @@ func TestListExperiments_Filtering(t *testing.T) {
 	assert.Nil(t, err)
 	assert.NotEqual(t, "", nextPageToken)
 	assert.Equal(t, expected, experiments)
-	assert.Equal(t, 3, count)
+	assert.Equal(t, int32(3), count)
 
 	// Next page should give experiment4.
 	opts, err = list.NewOptionsFromToken(nextPageToken, 2)
@@ -369,5 +369,5 @@ func TestListExperiments_Filtering(t *testing.T) {
 	// No more pages.
 	assert.Equal(t, "", nextPageToken)
 	assert.Equal(t, expected, experiments)
-	assert.Equal(t, 4, count)
+	assert.Equal(t, int32(4), count)
 }
