+++ backend/src/apiserver/storage/pipeline_store_test.go
@@ -66,7 +66,7 @@ func TestListPipelines_FilterOutNotReady(t *testing.T) {
 
 	assert.Nil(t, err)
 	assert.Equal(t, "", nextPageToken)
-	assert.Equal(t, 2, count)
+	assert.Equal(t, int32(2), count)
 	assert.Equal(t, pipelinesExpected, pipelines)
 }
 
@@ -100,7 +100,7 @@ func TestListPipelines_Pagination(t *testing.T) {
 	pipelines, count, nextPageToken, err := pipelineStore.ListPipelines(opts)
 	assert.Nil(t, err)
 	assert.NotEmpty(t, nextPageToken)
-	assert.Equal(t, 4, count)
+	assert.Equal(t, int32(4), count)
 	assert.Equal(t, pipelinesExpected, pipelines)
 
 	expectedPipeline2 := &model.Pipeline{
@@ -123,7 +123,7 @@ func TestListPipelines_Pagination(t *testing.T) {
 	pipelines, count, nextPageToken, err = pipelineStore.ListPipelines(opts)
 	assert.Nil(t, err)
 	assert.Empty(t, nextPageToken)
-	assert.Equal(t, 4, count)
+	assert.Equal(t, int32(4), count)
 	assert.Equal(t, pipelinesExpected2, pipelines)
 }
 
@@ -158,7 +158,7 @@ func TestListPipelines_Pagination_Descend(t *testing.T) {
 	pipelines, count, nextPageToken, err := pipelineStore.ListPipelines(opts)
 	assert.Nil(t, err)
 	assert.NotEmpty(t, nextPageToken)
-	assert.Equal(t, 4, count)
+	assert.Equal(t, int32(4), count)
 	assert.Equal(t, pipelinesExpected, pipelines)
 
 	expectedPipeline1 := &model.Pipeline{
@@ -180,7 +180,7 @@ func TestListPipelines_Pagination_Descend(t *testing.T) {
 	pipelines, count, nextPageToken, err = pipelineStore.ListPipelines(opts)
 	assert.Nil(t, err)
 	assert.Empty(t, nextPageToken)
-	assert.Equal(t, 4, count)
+	assert.Equal(t, int32(4), count)
 	assert.Equal(t, pipelinesExpected2, pipelines)
 }
 
@@ -202,7 +202,7 @@ func TestListPipelines_Pagination_LessThanPageSize(t *testing.T) {
 	pipelines, count, nextPageToken, err := pipelineStore.ListPipelines(opts)
 	assert.Nil(t, err)
 	assert.Equal(t, "", nextPageToken)
-	assert.Equal(t, 1, count)
+	assert.Equal(t, int32(1), count)
 	assert.Equal(t, pipelinesExpected, pipelines)
 }
